feat(web): migrate homepage and experience page content components to rsc
